requirements.txt
MyMySQL==1.0.2
python-dotenv==0.19.1
